Add files via upload
If putting /boot on a USB drive is too difficult for you, set a BIOS, hard drive and boot password. Furthermore you need to check the integrity of all files in /boot regularly, you can use my scripts for that.

calc_hash.sh --> Calculates, encrypts (using PGP) a SHA512 hash of /boot and stores it in a file called hash.enc. You should move it to a read-only device to prevent tampering.
integrity_check.sh --> Decrypts hash.enc and compares it to the current SHA512 hash of /boot.

Install gnustep-gui-runtime to get the fancy audio warning.